+++ NEWS.md
@@ -1,3 +1,15 @@
+# gibble 0.1.2
+
+* Fix gibble for PATH objects from silicate (when they are points). 
+
+* SpatialPoints was has object column, needed for silicate. 
+
+# gibble 0.1.0
+
+* Fixed multipoints for silicate, both sf and sp. 
+
+* Fixes for SpatialLines and SpatialMultiPoints, found in anglr. 
+
 # gibble 0.0.2
 
 * Added `PATH`, `sp` , `trip` methods for `gibble`.